Cleanups and corrections
